+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 - `writer` module is private now.
+- `Node::set_attribute` accepts only `Attribute` or tuple with attribute name and value now.
 
 ### Removed
 - `Node::attribute`. Use `node.attributes().get()` instead.
 - `Node::attribute_value`. Use `node.attributes().get_value()` instead.
 - `Node::has_attribute_with_value`.
+- `Node::set_link_attribute`. Use `Node::set_attribute` instead.
+- `Node::set_attribute_object`. Use `Node::set_attribute` instead.
 
 ## [0.6.0] - 2017-06-18
 ### Added